TestUF2406

+++ TestUF2406/src/main/java/com/example/TestUF2406/repository/BookRepository.java
@@ -1,10 +1,9 @@
 package com.example.TestUF2406.repository;
 
 import org.springframework.data.repository.CrudRepository;
-
 import com.example.TestUF2406.model.Book;
 
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 
 	
 }
